Fixed small bug when resizing got rid of unuseful or unused files.
